clean code

+++ Game/Game/CAniObject.h
@@ -6,7 +6,7 @@ class CAniObject :
 {
 protected:
     CTimer timer{ true, 1 };
-    float w, h;
+    int w, h;
 
 public:
     CAniObject(D3DXVECTOR2 position, float vx, float vy, int aniId, D3DXVECTOR2 scale = {1, 1}, long timeout = 2000, float gravity = 0.001f, int w = 8, int h = 8);
